Merged the last changes from the old HRP svn repository.
Removed the Change log part of the read me as it is probably not going to be possible to keep it up to date anyway (it's already missing a ton of changes I have done). Better to compile such a set of changes from the SVN revisions when releasing an official pack.