Polymost glowmaps added;  duke3d.def: remove irritating comment.
I'm not really happy with this solution but up to now only four files are affected by this rare Polymer/Polymost conflict. And so far I think it's better for maintenance not to introduce a textures_polymost folder.

duke3d.def: This is definetely not the place to split the DEF path between Polymer and Polymost.